pyproject.toml
[tool.poetry]
name = "ml_pipeline"
version = "0.1.0"
description = "ML Pipeline for Bank Marketing Prediction"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.0"
numpy = "^1.23.0"
scikit-learn = "^1.1.0"
xgboost = "^1.7.0"
python-dotenv = "^0.21.0"
hydra-core = "^1.3.0"
wandb = "^0.15.0"
pytest = "^7.3.0"
black = "^22.10.0"
isort = "^5.10.0"
flake8 = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
